% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAF.R
\name{BAF}
\alias{BAF}
\title{Bioaccumulation Factor (BAF) Calculation for Shellfish}
\usage{
BAF(n, mean_value = 44.9, sd_value = 20.93, lower = 1, upper = 100)
}
\arguments{
\item{n}{Number of iterations to generate BAF values.}

\item{mean_value}{Mean value for BAF (default = 44.9).}

\item{sd_value}{Standard deviation value for BAF (default = 20.93).}

\item{lower}{Lower bound for truncation in the distribution (default = 1).}

\item{upper}{Upper bound for truncation in the distribution (default = 100).}
}
\value{
A vector containing generated BAF values.
}
\description{
BAF (Bioaccumulation Factor) represents the tendency of shellfish to accumulate pathogens in their flesh.
}
\examples{
# Generate BAF values for 1000 iterations using default parameters
baf_values <- BAF(1000)

}
