% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{convertConcInLiterToMililiters}
\alias{convertConcInLiterToMililiters}
\title{convertConcInLiterToMililiters}
\usage{
convertConcInLiterToMililiters(effluent)
}
\arguments{
\item{effluent}{number, concentration per L}
}
\value{
effluent concentraion per mL
}
\description{
Deals with the fact that bug concentration is in L and we need it in mL
}
\examples{
convertConcInLiterToMililiters(1000)

}
