% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{create_percentile_df}
\alias{create_percentile_df}
\title{Create Percentile DataFrame from isIllNoro Matrix}
\usage{
create_percentile_df(isIllNoro)
}
\arguments{
\item{isIllNoro}{A matrix representing illness status data.}
}
\value{
A dataframe containing percentiles named similarly to those found in risk assessment reports.
}
\description{
This function takes the 'isIllNoro' matrix and generates a dataframe with percentiles commonly used in risk assessment reports.
The percentiles include values at 0.01, 0.025, 0.05, 0.1, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45, 0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.9, 0.95, 0.975, 0.99, 0.995, and 0.999.
}
\details{
The function calculates various percentiles based on the provided 'isIllNoro' matrix and creates a dataframe with these percentiles named accordingly.
}
\examples{
# Assuming isIllNoro is a matrix of illness status data
isIllNoro <- matrix(c(rep(c(0,1),50),rep(c(0,0,0,1),25)), ncol =2)
create_percentile_df(isIllNoro)
}
