% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{create_stats_df}
\alias{create_stats_df}
\title{Create Statistics DataFrame from isIllNoro Matrix}
\usage{
create_stats_df(isIllNoro)
}
\arguments{
\item{isIllNoro}{A matrix representing illness status data.}
}
\value{
A dataframe containing statistics named similarly to those found in risk assessment reports.
}
\description{
This function takes the 'isIllNoro' matrix and generates a dataframe with statistics commonly used in risk reports.
The statistics include Minimum, Maximum, Mean, Mode, Median, Standard Deviation, and Variance .
(Kurtosis and Skewness is not currently included in the statistics)
}
\details{
The function calculates various statistics (Minimum, Maximum, Mean, Mode, Median, Standard Deviation, Variance, Skewness, Kurtosis)
based on the provided 'isIllNoro' matrix and creates a dataframe with these statistics named accordingly.
The 'Mode' function used here refers to a custom function for estimating the mode of a vector.
}
\examples{
# Assuming isIllNoro is a matrix of illness status data
isIllNoro <- matrix(c(rep(c(0,1),50),rep(c(0,0,0,1),25)), ncol =2)
stats_df <- create_stats_df(isIllNoro)

}
\seealso{
\code{\link{Mode}}
}
