% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.R
\name{dhockey}
\alias{dhockey}
\title{dhockey}
\usage{
dhockey(x, X0, X50, X100, P = 0.95)
}
\arguments{
\item{x}{vector, bug concentration}

\item{X0}{number, minimum bug concentration}

\item{X50}{number, median big concentration}

\item{X100}{number, maximum bug concentration}

\item{P}{number (between 0 and 1) percentile for Xp}
}
\value{
vector, of probability densities
}
\description{
Estimates the density of the hockey distribution at any value of bug
concentration, x. Does This change Text
}
\examples{
dhockey(0,10,100,10000)

}
