% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{estEffluentConc}
\alias{estEffluentConc}
\title{Estimates concentrtion of effluent at the end of the pipe}
\usage{
estEffluentConc(n, X0, X50, X100, LRV, P = 0.95, MHF = 18.5)
}
\arguments{
\item{n}{number of iterations}

\item{X0}{minimum conc}

\item{X50}{median conc}

\item{X100}{maximum conc}

\item{LRV}{Treatment}

\item{P}{probability value relating to Xp (default is normally 0.95)}

\item{MHF}{Method Harmonisation Factor (default is normally 18.5)}
}
\value{
vector
}
\description{
Takes the estimate concentration of bugs in \strong{influent} (per L) and returns the
concentration per ml in the \strong{effluent} taking into account treatment and MHF.
This function is specifically for \strong{norovirus} and assumes Xp = 0.95
}
\examples{
estEffluentConc(5, 10, 1000, 1e6, 2, 0.95, 18.5)

}
