% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isIllNoro.R
\name{isIllNoro}
\alias{isIllNoro}
\title{Estimate Illness Status based on Infection Status}
\usage{
isIllNoro(isInfected, probIll = 0.6, propSuseptable = 0.74, mySeed = 87)
}
\arguments{
\item{isInfected}{A numeric matrix representing infection status (1 for infected, 0 for not infected).}

\item{probIll}{A numeric value representing the conditional probability of illness given infection status (default = 0.6).}

\item{propSuseptable}{A numeric value representing the propotion of population suseptiable to norovirus (default = 0.74)}

\item{mySeed}{An optional numeric seed for reproducibility in random number generation (default = 87).}
}
\value{
A numeric matrix representing the estimated illness status based on infection status and conditional probability.
}
\description{
This function estimates illness status based on a matrix of infection status ('isInfected') where 1 represents infected and 0 represents not infected.
The estimation relies on the conditional probability of illness given infection (P(ill|inf)). Note the default value used here differs from comes from
table A1.2 of the WHO guidance on "Quantitative Microbial Risk Assessment: Application for Water Safety Management", originating from Teunis et al. 2008.
The function also accounts for suseptability which is set at 0.74 of the population being suseptible to norovirus
}
\details{
This function estimates illness status based on the provided infection status matrix ('isInfected') and the given conditional probability of illness given infection.
It checks if the input infection status values are either 0 (not infected) or 1 (infected). If not, it throws an error.
The estimation relies on a random comparison using the conditional probability to determine illness status.
}
\examples{
infection_status <- matrix(c(1, 0, 1, 1, 0, 1), nrow = 2)
isIllNoro(infection_status)
isIllNoro(infection_status, probIll = 0.75, mySeed = 123)

}
\references{
McBride (various spreadsheets),
Cressey (2021) SCREENING QUANTITATIVE MICROBIAL RISK ASSESSMENT (QMRA): KAIKOHE WASTEWATER TREATMENT PLANT,
Teunis et al. (2008). Add citation/reference details.
}
\seealso{
\code{\link{runif}}
}
