% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInfected.R
\name{isInfected}
\alias{isInfected}
\title{Determine Infection Status based on Probabilities}
\usage{
isInfected(probInfection, mySeed = 98)
}
\arguments{
\item{probInfection}{A numeric matrix representing probabilities of infection for individuals.}

\item{mySeed}{An optional numeric seed for reproducibility in random number generation (default = 98).}
}
\value{
A numeric matrix indicating infection status (1 for infected, 0 for not infected).
}
\description{
This function determines whether an individual is infected or not based on a matrix of probabilities ('probInfection').
It checks if the input probabilities are within the range \link{0, 1} inclusive, and if not, throws an error.
If the probabilities are within the valid range, the function generates a random seed-based comparison to determine infection status.
}
\details{
This function evaluates whether individuals are infected based on the provided probabilities.
It checks if the input probabilities are within the acceptable range \link{0, 1}.
If the input probabilities are outside this range, the function throws an error.
Otherwise, it generates a seed-based random comparison to determine the infection status.
}
\examples{
prob_matrix <- matrix(runif(25), nrow = 5)
isInfected(prob_matrix)
isInfected(prob_matrix, mySeed = 123)

}
\seealso{
\code{\link{runif}}
}
