% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterEquiv.R
\name{mealSize}
\alias{mealSize}
\title{Estimate Shellfish Meal Size for Multiple Iterations}
\usage{
mealSize(n)
}
\arguments{
\item{n}{Number of iterations determining the number of groups (each of 100 people).}
}
\value{
A matrix representing the estimated meal sizes for each group (100 people) across 'n' iterations.
}
\description{
This function estimates the shellfish meal size in grams for 100 people across 'n' iterations.
Each iteration represents a group of 100 people, and 'n' determines the number of such iterations.
}
\examples{
# Estimate meal size for 500 iterations
meal_sizes <- mealSize(500)

}
