% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{methodHarmonisationFactor}
\alias{methodHarmonisationFactor}
\title{methodHarmonisationFactor}
\usage{
methodHarmonisationFactor(effluent, MHF)
}
\arguments{
\item{effluent}{vector, concentration of bugs in treated wastewater}

\item{MHF}{number, correcting factor (only used for Norovirus)}
}
\value{
correctedEffluent
}
\description{
Corrects concentration taking into account the measurement method
only used for norovirus
}
\examples{
# Correct concentration of Norovirus in treated wastewater using MHF 1.5
effluent_concentration <- c(10, 15, 20, 25)
corrected_concentration <- methodHarmonisationFactor(effluent_concentration, 1.5)

}
