% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{Mode}
\alias{Mode}
\title{Estimate Mode of a Vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{A numeric or categorical vector to estimate the mode from.}
}
\value{
The mode of the input vector 'x'.
}
\description{
This function estimates the mode of a vector based on the most frequently occurring value.
It identifies the value that appears most frequently (mode) in the input vector.
}
\details{
The function identifies the mode of the provided vector by finding the most frequently occurring value.
In cases where multiple values have the same maximum frequency, the function returns the first one encountered.
This implementation refers to the approach detailed at https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode.
}
\examples{
vector <- c(1, 2, 2, 3, 4, 4, 4, 5)
Mode(vector)

}
\references{
See https://stackoverflow.com/questions/2547402/how-to-find-the-statistical-mode
}
