% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.R
\name{phockey}
\alias{phockey}
\title{phockey}
\usage{
phockey(x, X0, X50, X100, P = 0.95)
}
\arguments{
\item{x}{vector, bug concentration}

\item{X0}{number, minimum bug concentration}

\item{X50}{number, median big concentration}

\item{X100}{number, maximum bug concentration}

\item{P}{number (between 0 and 1) percentile for Xp}
}
\value{
vector, cumulative probaility
}
\description{
Returns the cumulate probability for any value of bug.
}
\examples{
phockey(100, 10, 100, 1000)

}
