% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatedLogLogistic.R
\name{ploglosistic}
\alias{ploglosistic}
\title{cdf log logistic fucntion}
\usage{
ploglosistic(x, alpha, beta, gamm)
}
\arguments{
\item{x}{number}

\item{alpha}{number}

\item{beta}{number}

\item{gamm}{number}
}
\value{
number
}
\description{
estimates the cdf for the loglogistic
}
