% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.R
\name{qhockey}
\alias{qhockey}
\title{qhockey}
\usage{
qhockey(p, X0, X50, X100, P = 0.95)
}
\arguments{
\item{p}{vector, probability for the value of x to be calculated}

\item{X0}{number, minimum bug concentration}

\item{X50}{number, median big concentration}

\item{X100}{number, maximum bug concentration}

\item{P}{number (between 0 and 1) percentile for Xp}
}
\value{
vector of bug concentrations
}
\description{
Estimate bug concentration for any value of cumulative probability for the hockey stick distribution
}
\examples{
qhockey(0.5, 10, 100, 1000)

}
