% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hockey.R
\name{rhockey}
\alias{rhockey}
\title{rhockey}
\usage{
rhockey(n, X0, X50, X100, P = 0.95, mySeed = 123)
}
\arguments{
\item{n}{number, sample size}

\item{X0}{number, minimum bug concentration}

\item{X50}{number, median big concentration}

\item{X100}{number, maximum bug concentration}

\item{P}{number (between 0 and 1) percentile for Xp}

\item{mySeed}{number, random seed}
}
\value{
vector, bug concentrations
}
\description{
Draws a random (and reproducible) sample from the hockey stick distribution
}
\examples{
rhockey(10, 10, 100, 1000, P = 0.95, mySeed = 123)
}
