% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatedLogLogistic.R
\name{rloglosisticTrunc}
\alias{rloglosisticTrunc}
\title{cdf log logistic function}
\usage{
rloglosisticTrunc(
  n,
  alpha = 2.2046,
  beta = 75.072,
  gamm = -0.9032,
  minX = 5,
  maxX = 800,
  mySeed = 234
)
}
\arguments{
\item{n}{number the sample size}

\item{alpha}{number = 2.2046}

\item{beta}{number = 75.072}

\item{gamm}{number = -0.9032}

\item{minX}{number minimum value for the truncated value = 5g}

\item{maxX}{number maximum value for the truncated value = 800g}
}
\value{
number
}
\description{
Estimates the cdf for the truncated loglogistic function, which is used to
estimate n meal sizes.
}
