% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDose.R
\name{roundDose}
\alias{roundDose}
\title{Rounds a matrix of dose values to the nearest integers in a stochastic manner.}
\usage{
roundDose(rawDose, mySeed = 245)
}
\arguments{
\item{rawDose}{A numeric matrix representing the raw doses to be rounded.}

\item{mySeed}{An optional numeric seed for reproducibility in random number generation.}
}
\value{
A numeric matrix representing the rounded doses based on stochastic rounding.
}
\description{
This function performs stochastic rounding on a matrix of 'rawDose' values, providing a matrix of rounded doses based on randomness.
}
\examples{
matrix_input <- matrix(c(3.6, 7.2, 5.4, 6.7), nrow = 2)
roundDose(matrix_input)
roundDose(matrix_input, mySeed = 123)

}
