% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shellfishRiskModel.R
\name{shellfishRiskModel}
\alias{shellfishRiskModel}
\title{Shellfish Consumption Risk Model}
\usage{
shellfishRiskModel(n, X0, X50, X100, LRV, P, MHF, myCSVFile, mySite)
}
\arguments{
\item{n}{Integer specifying the number of Monte Carlo simulations.}

\item{X0}{Numeric specifying the minimum norovirus concentration in influent (gc/L).}

\item{X50}{Numeric specifying the median norovirus concentration in influent (gc/L).}

\item{X100}{Numeric specifying the maximum norovirus concentration in influent (gc/L).}

\item{LRV}{Numeric specifying the log reduction value for norovirus during sewage treatment.}

\item{P}{Numeric number (between 0 and 1) percentile for Xp (Our default values 0.95)}

\item{MHF}{Numeric Method Harmonisation Factor (Our default value is 18.5)}

\item{myCSVFile}{Character string specifying the name of the CSV file containing dilution data.}

\item{mySite}{Character string specifying the column name for site-specific dilution values.}
}
\value{
Numeric vector representing the total number of norovirus-infected cases in each simulation of 100 people.
}
\description{
This function calculates the risk of norovirus infection associated with shellfish consumption.
This model assumes norovirus is in a disagregated state
}
\details{
This function combines various sub-functions to estimate the risk of norovirus infection
based on shellfish consumption. It involves estimating effluent concentration, dilution factor,
diluted effluent concentration, water equivalent consumed, raw dose, rounded dose, probability of infection,
infection status, and the total number of infected cases in each simulation.
}
\examples{
\dontrun{
# Example usage
results <- shellFishRiskModel(1000, 0.5, 1.5, 2.0, 1.0, 0.2, 2.5, "your_data.csv", "yourSiteColumn")
}

}
\seealso{
\code{\link{estEffluentConc}}, \code{\link{estDilution}}, \code{\link{waterEquiv}},
\code{\link{roundDose}}, \code{\link{doseResponseNoro}}, \code{\link{isInfected}},
\code{\link{isIllNoro}}
}
