% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{solve_x}
\alias{solve_x}
\title{solve_x}
\usage{
solve_x(slope, height, area)
}
\arguments{
\item{slope}{number}

\item{height}{number}

\item{area}{number}
}
\value{
number
}
\description{
helper function, solves the quadratic equation
}
\examples{
solve_x(1,0,0.5)
}
