% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming.R
\name{swim_duration}
\alias{swim_duration}
\title{Estimate Swim Duration}
\usage{
swim_duration(n, min_duration, mode_duration, max_duration, mySeed = 876)
}
\arguments{
\item{n}{Number of swim durations to generate.}

\item{min_duration}{Minimum duration of a swim (typical values 0.2 hour).}

\item{mode_duration}{Mode duration of a swim (typical values 1 hour).}

\item{max_duration}{Maximum duration of a swim (typical value 4 hour).}
}
\value{
A matrix containing \code{n} swim durations.
}
\description{
The \code{swim_duration} function estimates swim duration based on provided parameters.
}
\details{
This function uses the rpert function from the mc2d package to generate swim durations.
The rpert function generates random samples from the truncated triangular distribution.
The parameters \code{min_duration}, \code{mode_duration}, and \code{max_duration} define
the shape of the distribution.
}
\examples{
\dontrun{
# Generate swim durations with default parameters
swim_durations <- swim_duration(10)

# Generate swim durations with custom parameters
custom_durations <- swim_duration(10, min_duration = 0.5, mode_duration = 1.5, max_duration = 3.0)
}

}
\seealso{
\code{\link[=rpert]{rpert}} function from the mc2d package.
}
