% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming.R
\name{swim_ingestion_rate}
\alias{swim_ingestion_rate}
\title{Ingestion rate (Swimmers primary contact)}
\usage{
swim_ingestion_rate(
  n,
  min_ingestion_rate,
  max_ingestion_rate,
  mean_ingestion_rate,
  sd_ingestion_rate,
  mySeed = 521
)
}
\arguments{
\item{n}{Number of iterations to generate BAF values.}

\item{min_ingestion_rate}{(typical default value= 5)}

\item{max_ingestion_rate}{(typical default value= 250).}

\item{mean_ingestion_rate}{(typical default value = 53)}

\item{sd_ingestion_rate}{(typical default value = 75)}
}
\value{
A vector containing generated BAF values.
}
\description{
Ingestion rate (mL/h) of water from primary contact with water. Takes the form of a truncated lognormal distribution
}
\examples{
# Generate BAF values for 1000 iterations using default parameters
baf_values <- BAF(1000)

}
