% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimRiskModel.R
\name{swimRiskModel}
\alias{swimRiskModel}
\title{Swim Risk Model}
\usage{
swimRiskModel(
  n,
  X0,
  X50,
  X100,
  LRV,
  P,
  MHF,
  myCSVFile,
  mySite,
  min_duration,
  mode_duration,
  max_duration,
  min_ingestion_rate,
  mean_ingestion_rate,
  sd_ingestion_rate,
  max_ingestion_rate
)
}
\arguments{
\item{n}{Number of simulations}

\item{X0}{Parameter for effluent concentration calculation}

\item{X50}{Parameter for effluent concentration calculation}

\item{X100}{Parameter for effluent concentration calculation}

\item{LRV}{Log Removal Value for effluent concentration calculation}

\item{P}{Parameter for effluent concentration calculation}

\item{MHF}{Method harmonisation factor for norovirus}

\item{myCSVFile}{Path to the CSV file containing water quality data}

\item{mySite}{Name of the site for water quality data}

\item{min_duration}{Minimum duration of swimming (minutes)}

\item{mode_duration}{Mode duration of swimming (minutes)}

\item{max_duration}{Maximum duration of swimming (minutes)}

\item{min_ingestion_rate}{Minimum ingestion rate (volume per minute)}

\item{mean_ingestion_rate}{Mean ingestion rate (volume per minute)}

\item{sd_ingestion_rate}{Standard deviation of ingestion rate (volume per minute)}

\item{max_ingestion_rate}{Maximum ingestion rate (volume per minute)}
}
\value{
A vector containing the number of infected individuals in each simulation.
}
\description{
This function calculates the risk of norovirus infection from #'swimming in water impacted by wastewater. It considers factors #'such as influent concentration, treattment (LRV), considering factors such as dilution, water consumption,
and dose-response for disagregated norovirus.
}
\details{
The function calculates the risk by estimating effluent concentration, dilution
factor, ingested volume, dose, and applying dose-response modeling.
}
\seealso{
\code{\link{estEffluentConc}}, \code{\link{estDilution}},
\code{\link{volume_ingested_swim}}, \code{\link{doseResponseNoro}},
\code{\link{isInfected}}, \code{\link{isIllNoro}}, \code{\link{roundDose}}
}
