% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\name{treatment}
\alias{treatment}
\title{treatment}
\usage{
treatment(influent, LRV)
}
\arguments{
\item{influent}{vector, concentration of bugs per L in influent}

\item{LRV}{numeric, treatment Log reduction value}
}
\value{
effluent concentration of bugs in effluent
}
\description{
Estimate of how how the
}
\examples{
treatment(1000,2)
}
