% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimming.R
\name{volume_ingested_swim}
\alias{volume_ingested_swim}
\title{Calculate Volume Ingested during Swim}
\usage{
volume_ingested_swim(
  n,
  min_duration = 0.2,
  mode_duration = 1,
  max_duration = 4,
  min_ingestion_rate = 5,
  max_ingestion_rate = 250,
  mean_ingestion_rate = 53,
  sd_ingestion_rate = 75
)
}
\arguments{
\item{n}{Number of swims to simulate.}

\item{min_duration}{Minimum duration of a swim (default = 0.2 h).}

\item{mode_duration}{Mode duration of a swim (default = 1 h).}

\item{max_duration}{Maximum duration of a swim (default = 4 h).}

\item{min_ingestion_rate}{Minimum ingestion rate during a swim (5 mL/h).}

\item{max_ingestion_rate}{Maximum ingestion rate during a swim (250 mL/h).}

\item{mean_ingestion_rate}{Mean ingestion rate during a swim (53 mL/h).}

\item{sd_ingestion_rate}{Standard deviation of ingestion rate during a swim (75 mL/h).}
}
\value{
A matrix containing the volume ingested during each of the \code{n} swims.
}
\description{
The \code{volume_ingested_swim} function calculates the volume ingested during a swim
based on swim duration and ingestion rate parameters.
}
\details{
This function utilizes the \code{swim_duration} and \code{swim_ingestion_rate} functions
to generate swim durations and ingestion rates, respectively, and calculates the volume ingested
during each swim by multiplying the corresponding duration and ingestion rate.
}
\examples{
\dontrun{
# Calculate volume ingested during swims with default parameters
volume_data <- volume_ingested_swim(10)

# Calculate volume ingested during swims with custom parameters
custom_volume_data <- volume_ingested_swim(10, min_duration = 0.5, mode_duration = 1.5,
                                           max_duration = 3.0, min_ingestion_rate = 10,
                                           max_ingestion_rate = 200, mean_ingestion_rate = 60,
                                           sd_ingestion_rate = 80)
}

}
\seealso{
\code{\link[=swim_duration]{swim_duration}} function for generating swim durations.
\code{\link[=swim_ingestion_rate]{swim_ingestion_rate}} function for generating ingestion rates.
}
