% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterEquiv.R
\name{waterEquiv}
\alias{waterEquiv}
\title{Convert Shellfish Meal Size to Water Volume Equivalent}
\usage{
waterEquiv(n)
}
\arguments{
\item{n}{Number of iterations determining the number of groups (each of 100 people).}
}
\value{
A matrix representing the estimated water volume equivalents (in ml) for each group (100 people) across 'n' iterations.
}
\description{
Estimates the equivalent volume of water in millilitres (ml) based on shellfish meal sizes for 100 people across 'n' iterations.
Each iteration represents a group of 100 people, exposed to the same Bioaccumulation Factor (BAF).
It assumes that BAF is described by truncate normal distribution with minimum and maximum values of 1 and 800g
mean = 49.9 and standard deviation = 20.03
It also assumes that meal size is described by a truncated log logistic (min 5g and maximum 800g) and
alpha = 	2.2046, beta = 75.072 and gamma =	-0.9032. These factors are hard coded
}
\examples{
# Estimate water volume equivalent for 500 iterations
water_equiv <- waterEquiv(500)

}
